+++ app/src/main/AndroidManifest.xml
@@ -68,13 +68,13 @@
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity"/>
 
-        <activity android:name=".all_activities.games.TicTacToeActivity"
+        <activity android:name=".all_activities.gamesActivities.TicTacToeActivity"
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity"/>
 
-        <service android:name=".model.helperClasses.TimerService" />
+        <service android:name=".model.helperClasses.timer.TimerService" />
         <service
-            android:name=".model.helperClasses.ForegroundService"
+            android:name=".model.helperClasses.foreground.ForegroundService"
             android:enabled="true"
             android:exported="false"
             android:foregroundServiceType="dataSync"/>
