+++ app/src/main/java/PozMaxPav/com/all_activities/ChildrenProfileActivity.java
@@ -5,9 +5,8 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
-import androidx.appcompat.app.AppCompatActivity;
 import PozMaxPav.com.R;
-import PozMaxPav.com.model.helperClasses.SharedPreferencesUtils;
+import PozMaxPav.com.model.helperClasses.sharedPreference.SharedPreferencesUtils;
 
 public class ChildrenProfileActivity extends BaseActivity {
