+++ app/src/main/java/PozMaxPav/com/model/helperClasses/sharedPreference/SharedPreferencesUtils.java
@@ -1,4 +1,4 @@
-package PozMaxPav.com.model.helperClasses;
+package PozMaxPav.com.model.helperClasses.sharedPreference;
 
 import android.content.Context;
 import android.content.SharedPreferences;
